+++ .changeset/great-mice-turn.md
@@ -1,0 +1,5 @@
+---
+"roo-cline": patch
+---
+
+Remove redundant zod schemas
